+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "turborand"
-version = "0.5.1"
+version = "0.6.0"
 edition = "2021"
 authors = ["Gonçalo Rica Pais da Silva <bluefinger@gmail.com>"]
 description = "A simple and fast random number generator"
